if the tweet is too long to be tweeted, just skip it
This can happen if the description text of the tweet contains a url that is *shorter* than the twitter link shortener. In this case the tweet will become unexpectedly longer, and the tweet will be longer than is allowed. Previously this would stop the bot from posting for a week while these tweets expired, now we just mark them as posted and move on